backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
openai>=1.12.0
google-generativeai>=0.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic[email]>=2.6.0
aiofiles>=23.2.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
sqlalchemy>=2.0.0
cryptography>=41.0.0

